wip: cartoon lipsync work + WSL setup script

+++ scripts/inference.py
@@ -116,6 +116,7 @@ def main(args):
                 output_vid_name = os.path.join(temp_dir, args.output_vid_name)
             output_vid_name_concat = os.path.join(temp_dir, output_basename + "_concat.mp4")
             
+            save_dir_full = None
             # Extract frames from source video
             if get_file_type(video_path) == "video":
                 save_dir_full = os.path.join(temp_dir, input_basename)
@@ -240,7 +241,8 @@ def main(args):
             shutil.rmtree(result_img_save_path)
             os.remove(temp_vid_path)
             
-            shutil.rmtree(save_dir_full)
+            if save_dir_full is not None and os.path.exists(save_dir_full):
+                shutil.rmtree(save_dir_full)
             if not args.saved_coord:
                 os.remove(crop_coord_save_path)
